requirements.txt
# -----------------------------
# Core Utilities
# -----------------------------
python-dotenv>=1.0.0
typing-extensions>=4.5.0
pydantic>=2.0.0
requests>=2.31.0

# -----------------------------
# LangChain Core
# -----------------------------
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0

# -----------------------------
# LangGraph (Workflows & Agents)
# -----------------------------
langgraph>=0.0.40
langgraph-supervisor>=0.0.8

# -----------------------------
# LLM Providers
# -----------------------------
google-generativeai>=0.5.0

# -----------------------------
# Vector Stores & Embeddings
# -----------------------------
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.40.0

# -----------------------------
# RAG Loaders
# -----------------------------
youtube-transcript-api>=0.6.2
beautifulsoup4>=4.12.0
lxml>=5.0.0

# -----------------------------
# Agents & Tooling
# -----------------------------
crewai>=0.30.0

# -----------------------------
# Web / API Frameworks
# -----------------------------
fastapi>=0.110.0
uvicorn>=0.29.0
starlette>=0.36.0
python-multipart>=0.0.9

# -----------------------------
# Visualization & Dev UX
# -----------------------------
rich>=13.7.0
ipython>=8.20.0
Pillow>=10.0.0